pyproject.toml
[project]
name = "ocde-p6"
version = "0.1.0"
description = ""
authors = [
    {name = "day811",email = "day811@laposte.net"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "panda (>=0.3.1,<0.4.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "nbconvert (>=7.16.6,<8.0.0)",
    "bentoml (>=1.4.26,<2.0.0)",
    "fastapi (>=0.119.1,<0.120.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
